+++ src/components/pages/home/index.tsx
@@ -6,6 +6,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 import { MenuItem } from '@/components/molecules/Menu';
 import BottomNavigation from '@/components/organisms/BottomNavigation';
+import EditGroupOverview from '@/components/organisms/EditGroupOverView';
 import GroupOverview from '@/components/organisms/GroupOverview';
 import Header from '@/components/organisms/Header';
 import HomeOverview from '@/components/organisms/HomeOverview';
@@ -18,6 +19,7 @@ const Home = () => {
 
   const userPathId = searchPatams.get('userId');
   const groupPathId = searchPatams.get('groupId');
+  const groupEditPathId = searchPatams.get('groupEditId');
 
   const [blockMenuItems] = useState<MenuItem[]>([
     {
@@ -42,13 +44,14 @@ const Home = () => {
           <HomeOverview />
         </div>
         <div className="pc">
-          {!userPathId && !groupPathId && (
+          {!userPathId && !groupPathId && !groupEditPathId && (
             <PcNavigation>
               ユーザーもしくはグループを選択してください
             </PcNavigation>
           )}
           {userPathId && <UsersOverview userId={userPathId} />}
           {groupPathId && <GroupOverview groupId={groupPathId} />}
+          {groupEditPathId && <EditGroupOverview groupId={groupEditPathId} />}
         </div>
       </main>
       <div className="sp">
